Add files via upload
This code performs CRUD operations over HTTP requests. Now, what has been implemented is Spring Security. This allows us to give access to certain endpoints based on the user roles. For example, roles like "GUEST", "EMPLOYEE", "MANAGER", "ADMIN". These roles are assigned to the specific user in which will be given certain or even all of the endpoints depending on their role. 

In the code we define a "SecurityFilterChain" that filters and sets the restricting access points to different HTTP methods based upon their role. In this demonstration we only used the "{noop}" which is just plain text password. However there is also "{bcrypt}" in which can be implemented to ensure more secure password storage.